fix(build): Fixed build issues and cleaned up old move code
