--- src/features/campaign-gm-screen/components/SettingsSection/index.ts
@@ -1,1 +1,0 @@
-export * from "./SettingsSection";
