--- src/types/Settings.type.ts
@@ -1,5 +1,0 @@
-import { Move } from "./Moves.type";
-
-export interface Settings {
-  customMoves: Move[];
-}
